Refactor PolicyEngine from CLI to Core package
Moves PolicyEngine, TOML loading, and configuration logic from @google/gemini-cli to @google/gemini-cli-core. This enables other packages to use the policy system while maintaining identical functionality for the CLI.
