+++ packages/core/src/config/config.test.ts
@@ -7,11 +7,9 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import type { Mock } from 'vitest';
 import type { ConfigParameters, SandboxConfig } from './config.js';
-import {
-  Config,
-  ApprovalMode,
-  DEFAULT_FILE_FILTERING_OPTIONS,
-} from './config.js';
+import { Config, DEFAULT_FILE_FILTERING_OPTIONS } from './config.js';
+import { ApprovalMode } from '../policy/types.js';
+
 import * as path from 'node:path';
 import { setGeminiMdFilename as mockSetGeminiMdFilename } from '../tools/memoryTool.js';
 import {
