+++ packages/core/src/tools/web-fetch.ts
@@ -19,7 +19,9 @@ import type { MessageBus } from '../confirmation-bus/message-bus.js';
 import { ToolErrorType } from './tool-error.js';
 import { getErrorMessage } from '../utils/errors.js';
 import type { Config } from '../config/config.js';
-import { ApprovalMode, DEFAULT_GEMINI_FLASH_MODEL } from '../config/config.js';
+import { DEFAULT_GEMINI_FLASH_MODEL } from '../config/config.js';
+import { ApprovalMode } from '../policy/types.js';
+
 import { getResponseText } from '../utils/partUtils.js';
 import {
   fetchWithTimeout,
